plate遷移ページの実装など
