--- app/models/address.rb
@@ -1,11 +1,0 @@
-class Address < ApplicationRecord
-
-  belongs_to :user
-  belong_to :plate
-  has_many :plates, through: warehouses
-
-  with_options presence: true do
-    validates :warehouse_number, :warehouse_colum,
-    :house_number, :number_of_stage
-  end
-end
