+++ app/models/warehouse.rb
@@ -1,4 +1,10 @@
 class Warehouse < ApplicationRecord
-  belong_to :plate
-  belong_to :address
+
+  belongs_to :user
+  belongs_to :plate
+
+  with_options presence: true do
+    validates :warehouse_number, :warehouse_colum,
+    :house_number, :number_of_stage
+  end
 end
