+++ app/views/home/top.html.erb
@@ -3,7 +3,6 @@
 
   <div class="user-contents">
   <% if user_signed_in? %>
-      <%= link_to current_user.name, "#", class: "user-name" %>
       <h2 class="service-explain">ログアウトは
       <%= link_to "こちら", destroy_user_session_path, method: :delete, class: "logout" %>
       </h2>
@@ -20,8 +19,8 @@
   <%#以下はログイン状態じゃないと表示されない%>
   <% if user_signed_in? %>
   <div class="item-contents">
-    <h2 class='service-explain'>商品の登録は
-      <a><%= link_to "こちら", new_plate_path, class: "item-link" %></a>
+    <ul class='service-explain'>
+      <li><%= link_to "プレート一覧", plates_path, class: "item-link" %></li>
     </h2>
   </div>
 
@@ -38,4 +37,18 @@
   </div>
   <% end %>
 </div>
+
+<table border="1">
+<tr>
+  <th>登録者</th><th>型</th><th>厚み</th><th>硬さ</th><th>色</th><th>高さ</th><th>長さ</th><th>枚数</th><th>倉庫名</th><th>列</th><th>番地</th><th>段数</th>
+</tr>
+<% @warehouse.each do |warehouse| %>
+<tr>
+  <td><%= warehouse.user.name %></td>
+  <td><%= warehouse.mold_id.name %></td>
+  <td><%= warehouse.thickness.name %></td><td><%= warehouse.hardness.name %></td><td><%= warehouse.plate_color.name %></td><td><%= warehouse.plate_width.name %></td><td><%= warehouse.plate_length.name %></td><td><%= warehouse.number_of_sheets %></td>
+</tr>
+</table>
+<% end %>
+
 <%= render "shared/footer" %>
