+++ app/views/plates/index.html.erb
@@ -1,0 +1,21 @@
+<%= render "shared/header" %>
+<h2 class="plate-registration">
+  <a><%= link_to "プレート新規登録", new_plate_path, class: "new-plate-link" %>
+</h2>
+
+<div class="plate-index-text">
+  <h2 class="plate-index-title">プレート登録一覧</h2>
+</div>
+<ul class="plate-contents">
+  <% @plate.each do |plate| %>
+  <li><%= plate.mold.name %>-<%= plate.thickness.name %>-<%= plate.hardness.name %> <%= plate.plate_color.name %> 
+  <%= plate.plate_width.name %>x<%= plate.plate_length.name %> <%= plate.number_of_sheets %>枚 
+  <%= link_to "倉庫へ登録", new_plate_warehouse_path(plate.id), class: "link-btn" %>/
+  <%= link_to "編集", edit_plate_path(plate.id), method: :get, class: "link-btn" %>/
+  <%= link_to "削除", plate_path(plate.id), method: :delete, class: "link-btn" %>
+  </li>
+  <% end %>
+</ul>
+
+<%= link_to "topページへ戻る", "/", class: "link_btn" %>
+<%= render "shared/footer" %>
