+++ app/views/warehouses/new.html.erb
@@ -1,0 +1,35 @@
+<%= render "shared/header" %>
+
+<div class="plate-Registration-main">
+  <h3 class="plate-Registration-title">登録したい情報の入力して下さい</h3>
+  <h4 class="plate-Registration-sub-title">全ての項目の情報は必須です</h4>
+
+  <h3 class="plate-text">プレートの情報を入力してください</h3>
+  <%= form_with model:@warehouse, url:plate_warehouses_path, local: true do |f| %>
+  <div class="weight-bold-text">倉庫ナンバー
+    <span class="indispensable">必須</span>
+  </div>
+  <%= f.text_field :warehouse_number, class:"warehouse_number_input", id:"warehouse_number", placeholder:"例）1" %>
+
+  <div class="weight-bold-text">列
+        <span class="indispensable">必須</span>
+      </div>
+    <%= f.text_field :warehouse_colum, class:"warehouse_colum_input", id:"warehouse_colum", placeholder:"例）1" %>
+
+  <div class="weight-bold-text">番地
+        <span class="indispensable">必須</span>
+      </div>
+    <%= f.text_field :house_number, class:"house_number_input", id:"house_number",placeholder:"例）1" %>
+
+  <div class="weight-bold-text">段数
+        <span class="indispensable">必須</span>
+      </div>
+    <%= f.text_field :number_of_stage, class:"number_of_stage_input", id:"number_of_stage", placeholder:"例）1" %>
+
+      <div class="Registration-btn-contents">
+        <%= f.submit "登録する" ,class:"Registration-btn" %>
+      </div>
+  <% end %>
+</div>
+
+<%= render "shared/footer" %>
