+++ db/schema.rb
@@ -10,18 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_09_24_142929) do
-
-  create_table "addresses", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
-    t.string "warehouse_number", null: false
-    t.string "warehouse_colum", null: false
-    t.integer "house_number", null: false
-    t.integer "number_of_stage", null: false
-    t.bigint "user_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["user_id"], name: "index_addresses_on_user_id"
-  end
+ActiveRecord::Schema.define(version: 2020_09_25_111610) do
 
   create_table "plates", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.integer "mold_id", null: false
@@ -40,28 +29,29 @@ ActiveRecord::Schema.define(version: 2020_09_24_142929) do
   create_table "users", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "name", null: false
     t.string "employee_number", null: false
-    t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
 
   create_table "warehouses", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
+    t.string "warehouse_number", null: false
+    t.string "warehouse_colum", null: false
+    t.integer "house_number", null: false
+    t.integer "number_of_stage", null: false
+    t.bigint "user_id", null: false
     t.bigint "plate_id", null: false
-    t.bigint "address_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.index ["address_id"], name: "index_warehouses_on_address_id"
     t.index ["plate_id"], name: "index_warehouses_on_plate_id"
+    t.index ["user_id"], name: "index_warehouses_on_user_id"
   end
 
-  add_foreign_key "addresses", "users"
   add_foreign_key "plates", "users"
-  add_foreign_key "warehouses", "addresses"
   add_foreign_key "warehouses", "plates"
+  add_foreign_key "warehouses", "users"
 end
